rename dynamic writing routes
